Add files via upload
This project is intended for beginners who want to learn how to create a GUI application in java using swing and Awt libraries. Swing and Awt are two of the most popular and widely used GUI toolkits for java, which provide various components and layouts for creating user interfaces. Swing is a newer and more advanced library than Awt, which offers more features and customization options. However, swing still relies on Awt for some of its functionality, such as event handling and graphics. Therefore, this project uses both swing and Awt components to create a functional and attractive quiz application.